Move api_client_instance fixture to test_api_client, remove unused fixtures
